requirements.txt
PyJWT
flask
SQLAlchemy
flask-cors
flask-migrate